seperated SetupUI
